requirements.txt
google-auth
google-auth-oauthlib==1.2.0 
google-api-python-client
PyYAML>=6.0.1
openai-whisper  # For local whisper transcription
ffmpeg-python   # For audio processing
openai>=1.10.0  # Updated OpenAI API package with audio transcription support
tiktoken>=0.5.0 # For token counting
psycopg2-binary
python-dotenv